get docling from settings
